Fix overwriting nginx headers when specifying headers in nested blocks.
If you add an add_header directive all add_headers in all containing
blocks are erased. This is documented with nginx but is a startling
behavior. I had added some headers for cache control and they wiped
headers for literally everythign else. This workaround is better than
most stuff I found online.  It's not as flexible, aka it may be noisy in
many other cases, but in this case I find it satisfying. Elegant even.