Remove try files from nginx location block for old site.
The nested location block for the "v1" of the site required the
resources linked from the index.html file to be prefixed with /v1 as well. The choice of
server prefix shouldn't bleed into the html for somethign so simple.
Part of the issue was that there is a known bug with alias and try_files
in the a location block. So I fixed that.

Now we don't have to change links in the index file if we choose to
change the path where this old version of the site is hosted. Also,
we'll most likely use the same pattern for the next "migration".