Reorder projects and reowrd their captions.
There are some design inconsistencies taht occur when the captions are
of different lengths. I wanted to fix those but updating the copy needed
to be done as well so why not the proverbial Fix 2 bugs by cancelling
them out.
Also, there were a bunch of typos in the captions lol. and they made no
sense.